+++ api/konnect/v1alpha2/konnect_apiauthconfiguration_ref.go
@@ -8,6 +8,14 @@ type KonnectAPIAuthConfigurationRef struct {
 	// +kubebuilder:validation:MinLength=1
 	// +kubebuilder:validation:MaxLength=253
 	Name string `json:"name"`
+
+	// Namespace is the namespace of the KonnectAPIAuthConfiguration resource.
+	// If not specified, defaults to the resource namespace.
+	//
+	// +optional
+	// +kubebuilder:validation:MinLength=1
+	// +kubebuilder:validation:MaxLength=253
+	Namespace *string `json:"namespace,omitempty"`
 }
 
 // ControlPlaneKonnectAPIAuthConfigurationRef is a reference to a KonnectAPIAuthConfiguration resource
@@ -21,7 +29,7 @@ type ControlPlaneKonnectAPIAuthConfigurationRef struct {
 	Name string `json:"name"`
 
 	// Namespace is the namespace of the KonnectAPIAuthConfiguration resource.
-	// If not specified, defaults to the same namespace as the KonnectConfiguration resource.
+	// If not specified, defaults to the resource namespace.
 	//
 	// +optional
 	// +kubebuilder:validation:MinLength=1