+++ crd-from-oas/pkg/generator/reconciler.go
@@ -120,7 +120,13 @@ func {{.EntityNameLowerCamel}}APIAuthConfigurationRef(object client.Object) []st
 		return nil
 	}
 
-	return []string{ent.GetNamespace() + "/" + ent.Spec.KonnectConfiguration.APIAuthConfigurationRef.Name}
+	namespace := ent.GetNamespace()
+	if ent.Spec.KonnectConfiguration.APIAuthConfigurationRef.Namespace != nil &&
+		*ent.Spec.KonnectConfiguration.APIAuthConfigurationRef.Namespace != "" {
+		namespace = *ent.Spec.KonnectConfiguration.APIAuthConfigurationRef.Namespace
+	}
+
+	return []string{namespace + "/" + ent.Spec.KonnectConfiguration.APIAuthConfigurationRef.Name}
 }
 `